pyproject.toml
[project]
name = "e2e-data-engineering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "kafka-python>=2.2.15",
  "pandas>=2.3.3",
  "plotly>=6.3.1",
  "psycopg2-binary>=2.9.10",
  "pyhive>=0.7.0",
  "pyspark>=3.5.7",
  "setuptools>=80.9.0",
  "streamlit>=1.50.0",
  "thrift>=0.22.0",
  "thrift-sasl>=0.4.3",
]
